blocks coordinates bug fixed

+++ include/chunk_manager.h
@@ -34,7 +34,7 @@ public:
 	void updateChunks(glm::vec3 playerPosition);
 	void unloadChunks(glm::vec3 playerPosition);
 
-	Chunk* getChunk(int x, int y, int z) const;
+	Chunk* getChunk(float x, float y, float z) const;
 	std::unordered_map<glm::ivec3, Chunk*> getChunks() const { return m_chunks; }
 
 	void updateChunk(Chunk* chunk);
